Update for API 14

+++ PartyIcons/Runtime/NameplateUpdater.cs
@@ -349,7 +349,7 @@ public sealed class NameplateUpdater : IDisposable
         imageNode->SetHeight(32);
 
         imageNode->WrapMode = 1;
-        imageNode->Flags = (byte)ImageNodeFlags.AutoFit;
+        imageNode->Flags = ImageNodeFlags.AutoFit;
         imageNode->LoadIconTexture(60071, 0);
 
         return imageNode;
